feature: Добавить логику аутентификации и авторизации, валидацию при регистрации, обработку ошибок при API запросах

+++ controllers/cards.js
@@ -1,53 +1,62 @@
 const Card = require('../models/card');
 const {
-  SERVER_ERROR_CODE,
-  BAD_REQUEST_CODE,
-  NOT_FOUND_CODE,
-  SERVER_ERROR_MESSAGE,
   INVALID_DATA_MESSAGE,
   NOT_FOUND_CARD_ID_MESSAGE,
   CAST_ERROR_MESSAGE,
+  FORBIDDEN_ERROR_MESSAGE,
 } = require('../utils/constants');
 const NotFoundError = require('../errors/NotFoundError');
+const BadRequestError = require('../errors/BadRequestError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ cards }))
-    .catch(() => {
-      res.status(SERVER_ERROR_CODE).send({ message: SERVER_ERROR_MESSAGE });
+    .catch((err) => {
+      next(err);
     });
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   Card.create({ name, link, owner })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST_CODE).send({ message: INVALID_DATA_MESSAGE });
+        next(new BadRequestError(INVALID_DATA_MESSAGE));
       } else {
-        res.status(SERVER_ERROR_CODE).send({ message: SERVER_ERROR_MESSAGE });
+        next(err);
       }
     });
 };
 
-module.exports.removeCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.removeCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .orFail(new NotFoundError(NOT_FOUND_CARD_ID_MESSAGE))
-    .then((user) => res.send({ data: user }))
+    .then((card) => {
+      const user = String(req.user._id);
+      const cardOwner = String(card.owner);
+
+      if (user === cardOwner) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((removedCard) => res.send(removedCard));
+      } else {
+        next(new ForbiddenError(FORBIDDEN_ERROR_MESSAGE));
+      }
+    })
     .catch((err) => {
       if (err.name === 'NotFound') {
-        res.status(NOT_FOUND_CODE).send({ message: NOT_FOUND_CARD_ID_MESSAGE });
+        next(new NotFoundError(NOT_FOUND_CARD_ID_MESSAGE));
       } else if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_CODE).send({ message: CAST_ERROR_MESSAGE });
+        next(new BadRequestError(CAST_ERROR_MESSAGE));
       } else {
-        res.status(SERVER_ERROR_CODE).send({ message: SERVER_ERROR_MESSAGE });
+        next(err);
       }
     });
 };
 
-module.exports.setLike = (req, res) => {
+module.exports.setLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -57,16 +66,16 @@ module.exports.setLike = (req, res) => {
     .then((cards) => res.send(cards))
     .catch((err) => {
       if (err.name === 'NotFound') {
-        res.status(NOT_FOUND_CODE).send({ message: NOT_FOUND_CARD_ID_MESSAGE });
+        next(new NotFoundError(NOT_FOUND_CARD_ID_MESSAGE));
       } else if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_CODE).send({ message: CAST_ERROR_MESSAGE });
+        next(new BadRequestError(CAST_ERROR_MESSAGE));
       } else {
-        res.status(SERVER_ERROR_CODE).send({ message: SERVER_ERROR_MESSAGE });
+        next(err);
       }
     });
 };
 
-module.exports.removeLike = (req, res) => {
+module.exports.removeLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -76,11 +85,11 @@ module.exports.removeLike = (req, res) => {
     .then((cards) => res.send(cards))
     .catch((err) => {
       if (err.name === 'NotFound') {
-        res.status(NOT_FOUND_CODE).send({ message: NOT_FOUND_CARD_ID_MESSAGE });
+        next(new NotFoundError(NOT_FOUND_CARD_ID_MESSAGE));
       } else if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_CODE).send({ message: CAST_ERROR_MESSAGE });
+        next(new BadRequestError(CAST_ERROR_MESSAGE));
       } else {
-        res.status(SERVER_ERROR_CODE).send({ message: SERVER_ERROR_MESSAGE });
+        next(err);
       }
     });
 };
